relocated runtime_memory_image to intermediate_representation
