+++ src/cwe_checker_lib/src/analysis/string_abstraction/context/trait_impls/tests.rs
@@ -12,8 +12,9 @@ use crate::{
             tests::mock_project_with_intraprocedural_control_flow, tests::Setup as ProjectSetup,
         },
     },
-    intermediate_representation::{Bitvector, Blk, ByteSize, ExternSymbol, Jmp, Tid, Variable},
-    utils::binary::RuntimeMemoryImage,
+    intermediate_representation::{
+        Bitvector, Blk, ByteSize, ExternSymbol, Jmp, RuntimeMemoryImage, Tid, Variable,
+    },
 };
 
 #[test]
